Remove redundant int conversion on round()

+++ lib/matplotlib/patches.py
@@ -2318,9 +2318,13 @@ class BoxStyle(_Style):
 
             # the sizes of the vertical and horizontal sawtooth are
             # separately adjusted to fit the given box size.
-            dsx_n = int(round((width - tooth_size) / (tooth_size * 2))) * 2
+            #
+            # Due to a bug in older numpy versions where round(np.float64)
+            # returns a float instead of an int, convert to float before round:
+            # (see gh#11810, adapted in gh#19321)
+            dsx_n = round(float((width - tooth_size) / (tooth_size * 2))) * 2
             dsx = (width - tooth_size) / dsx_n
-            dsy_n = int(round((height - tooth_size) / (tooth_size * 2))) * 2
+            dsy_n = round(float((height - tooth_size) / (tooth_size * 2))) * 2
             dsy = (height - tooth_size) / dsy_n
 
             x0, y0 = x0 - pad + tooth_size2, y0 - pad + tooth_size2
